perf(competition): T43 E16 vendor byte diet, view-returned state, generic back to E13
- enflame: cat/out stay in input dtype, state-copy block removed,
  new_state returned as x_cat tail-slice view (E15 traceback pinned
  the compile break to the separate state-copy kernel only)
- kunlunxin: pack built directly as channel-contiguous [B,L+S,D] in
  input dtype via one cat-of-permutes; affine kernel bytes frozen;
  state-copy kernel dropped for a tail-slice view
- ascend: single variable - cat in input dtype (E15 hoisting kept)
- generic: revert unsubmitted rolling bytes to platform-proven E13
- tests: chained-update reuse, no-storage-alias, mixed input dtypes

+++ src/flaggems_sglang/runtime/backend/_ascend/ops/causal_conv1d_update.py
@@ -137,8 +137,12 @@ def causal_conv1d_update(x, conv_state, weight, bias=None, activation="silu"):
     if squeeze_out:
         x = x.unsqueeze(-1)
     orig_dtype = x.dtype
-    # Concatenate state+x along time axis (data layout, not computation)
-    x_cat = torch.cat([conv_state.float(), x.float()], dim=-1).contiguous()
+    # Concatenate state+x along the time axis in the input dtype (data
+    # layout, not computation); the kernel's .to(tl.float32) loads are
+    # the exact cast the reference applies first, so no fp32 staging
+    # pass is needed. Mixed dtypes promote like the reference's
+    # .float() pair.
+    x_cat = torch.cat([conv_state, x], dim=-1).contiguous()
     wt = weight.t().contiguous().float()  # [W, D]
     if bias is not None:
         bias = bias.contiguous().float()
